+++ index.html
@@ -15,10 +15,8 @@
       <h1 class="promo__title visually-hidden">Keksobooking. Кексы по соседству</h1>
       <img src="img/keksobooking.svg" alt="Keksobooking. Кексы по соседству" width="215" height="45">
     </div>
-
     <!-- Карта объявлений -->
     <section class="map">
-
       <!-- Блок для вставки карты -->
       <div class="map__canvas" id="map-canvas"></div>
 
@@ -214,6 +212,12 @@
         <img src="" class="popup__photo" width="45" height="40" alt="Фотография жилья">
       </div>
     </article>
+  </template>
+
+  <template id="error-message">
+    <div class="error-message">
+      <p class="error-message__text">Ошибка</p>
+    </div>
   </template>
 
   <!-- Сообщение об успешном создании объявления -->
