+++ js/main.js
@@ -1,68 +1,32 @@
-import {createAd} from './data.js';
-import {createSimilarAdElement} from './popup.js';
-import {deactivateForm, activateForm} from './form.js';
+// import {createAd} from './data.js';
 
-const SIMILAR_OBJECT_COUNT = 10;
-const LAT_DEFAULT = 35.6894;
-const LNG_DEFAULT = 139.692;
-let userNumber = 1;
-const similarAds = new Array(SIMILAR_OBJECT_COUNT).fill(null).map(() => createAd(userNumber++));
+import {deactivateForm} from './form.js';
+import {onSuccess, onError} from './user-form.js';
+
+// const SIMILAR_OBJECT_COUNT = 10;
+
+// let userNumber = 1;
+// const similarAds = new Array(SIMILAR_OBJECT_COUNT).fill(null).map(() => createAd(userNumber++));
+
+// console.log(similarAds.json());
 
 deactivateForm();
-document.querySelector('#address').value = `${LAT_DEFAULT.toFixed(5)}, ${LNG_DEFAULT.toFixed(5)}`;
 
-const map = L.map('map-canvas').on('load', () => {
-  activateForm();
-}).setView({
-  lat: LAT_DEFAULT,
-  lng: LNG_DEFAULT,
-}, 10);
-
-const mainPinIcon = L.icon({
-  iconUrl: '../img/main-pin.svg',
-  iconSize: [52, 52],
-  iconAnchor: [26, 52],
-});
-
-const pinIcon = L.icon({
-  iconUrl: '../img/pin.svg',
-  iconSize: [40, 40],
-  iconAnchor: [20, 40],
-});
-
-const mainPinMarker = L.marker(
+fetch(
+  'https://23.javascript.pages.academy/keksobooking/data',
   {
-    lat: LAT_DEFAULT,
-    lng: LNG_DEFAULT,
+    method: 'GET',
+    credentials: 'same-origin',
   },
-  {
-    draggable: true,
-    icon: mainPinIcon,
-  },
-);
-mainPinMarker.addTo(map);
+)
+  .then((response) => {
+    if (response.ok) {
+      return response.json();
+    }
 
-L.tileLayer(
-  'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
-  {
-    attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors | Icons made by <a href="https://www.freepik.com" title="Freepik">Freepik</a> from <a href="https://www.flaticon.com/" title="Flaticon">www.flaticon.com</a>',
-  },
-).addTo(map);
-
-mainPinMarker.on('moveend', () => {
-  document.querySelector('#address').value = `${mainPinMarker.getLatLng().lat.toFixed(5)}, ${mainPinMarker.getLatLng().lng.toFixed(5)}`;
-});
-
-similarAds.forEach(({author, offer, location}) => {
-  const lat = location.lat;
-  const lng = location.lng;
-  const marker = L.marker({
-    lat,
-    lng,
-  },
-  {
-    draggable: true,
-    icon: pinIcon,
+    throw new Error(`${response.status} ${response.statusText}`);
+  }).then((data) => {
+    onSuccess(data);
+  }).catch((err) => {
+    onError(err);
   });
-  marker.addTo(map).bindPopup(createSimilarAdElement(author, offer));
-});
